Refactor toastify notifications on lists, and rename errorPage's folder
