+++ mycodereactserver/src/Pages/CodeRegister.jsx
@@ -6,7 +6,7 @@ import { codeRegistration } from "../Services/Backend.Endpoints";
 import { toast } from "react-toastify";
 import CodeForm from "../Components/Forms/CodeForm";
 import Loading from "../Components/Loading/Loading";
-import ErrorPage from "./Service/ErrorPage";
+import ErrorPage from "./Services/ErrorPage";
 import "react-toastify/dist/ReactToastify.css";
 
 const CodeRegister = () => {
